+++ Si.GameEngine/Core/TickControllers/EnemyDronesSpriteTickController.cs
@@ -18,7 +18,7 @@ namespace Si.GameEngine.Core.TickControllers
             _gameEngine = gameEngine;
         }
 
-        public override void ExecuteWorldClockTick(double epoch, SiVector displacementVector)
+        public override void ExecuteWorldClockTick(float epoch, SiVector displacementVector)
         {
             foreach (var enemy in Visible().Where(o => o.IsDrone == true))
             {
